pyproject.toml
[tool.poetry]
name = "django_ghostpost"
version = "0.1.0"
description = ""
authors = ["Drew Radcliff <sirampr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
